Move Distribution trait and associates to sub-module
